# Vectorize RandomTranslation