--- a/keras_cv/layers/preprocessing/random_translation_test.py
+++ b/keras_cv/layers/preprocessing/random_translation_test.py
@@ -192,6 +192,17 @@ class RandomTranslationTest(tf.test.TestCase, parameterized.TestCase):
         actual_output = layer(input_images, training=False)
         self.assertAllClose(expected_output, actual_output)
 
+    def test_random_translation_on_batched_images_independently(self):
+        image = tf.random.uniform(shape=(100, 100, 3))
+        input_images = tf.stack([image, image], axis=0)
+
+        layer = preprocessing.RandomTranslation(
+            height_factor=0.5, width_factor=0.5
+        )
+
+        results = layer(input_images)
+        self.assertNotAllClose(results[0], results[1])
+
     def test_config_with_custom_name(self):
         layer = preprocessing.RandomTranslation(0.5, 0.6, name="image_preproc")
         config = layer.get_config()