# What does this PR do?

Related to #291 

Benchmark results:

||Mode|n=100|n=200|n=500|n=1000|
|-|-|-|-|-|-|
|OldRandomTranslation|Eager|0.77s|1.69s|7.84s|30.13s|
|OldRandomTranslation|Graph|0.33s|1.19s|7.41s|29.92s|
|RandomTranslation|Eager|0.0072s|0.0069s|0.0087s|0.0119s|
|RandomTranslation|Graph|0.0014s|0.0019s|0.0042s|0.0077s|

![comparison](https://user-images.githubusercontent.com/20734616/224237502-b1b3f136-933e-4419-b8ab-56ee0708f24e.png)

OldRandomTranslation is too slow that I need to set num_images = [100, 200, 500, 1000] for benchmarks

Both implementations fail to run with XLA due to lacking ImageProjectiveTransformV3 support.
https://github.com/tensorflow/tensorflow/issues/55194

Similar to #1435, I add a test to check random translation is applied for each image.

```python
    def test_random_translation_on_batched_images_independently(self):
        image = tf.random.uniform(shape=(100, 100, 3))
        input_images = tf.stack([image, image], axis=0)

        layer = preprocessing.RandomTranslation(
            height_factor=0.5, width_factor=0.5
        )

        results = layer(input_images)
        self.assertNotAllClose(results[0], results[1])
```

Also, I found that `RandomTranslation` is missing in `keras_cv/layers/preprocessing/ragged_image_test.py` and `keras_cv/layers/preprocessing/with_labels_test.py`. I added it to both tests and they passed on my local env.

```python
    (
        "RandomTranslation",
        layers.RandomTranslation,
        {"height_factor": 0.5, "width_factor": 0.5},
    ),
```

## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [X] Did you read the [contributor guideline](https://github.com/keras-team/keras-cv/blob/master/.github/CONTRIBUTING.md),
      Pull Request section?
- [X] Was this discussed/approved via a Github issue? Please add a link
      to it if that's the case.
- [X] Did you write any new necessary tests?
- [ ] If this adds a new model, can you run a few training steps on TPU in Colab to ensure that no XLA incompatible OP are used?

## Who can review?
@LukeWood @ianstenbit